Struct bumpalo::boxed::Box

#[repr(transparent)]
pub struct Box<'a, T: ?Sized>(/* private fields */);

An owned pointer to a bump-allocated T value, that runs Drop implementations.

See the module-level documentation for more details.

Implementations

impl<'a, T> Box<'a, T>

pub fn new_in(x: T, a: &'a Bump) -> Box<'a, T>

Allocates memory on the heap and then places x into it.

This doesn’t actually allocate if T is zero-sized.

Examples

use bumpalo::{Bump, boxed::Box};

let b = Bump::new();

let five = Box::new_in(5, &b);

pub fn pin_in(x: T, a: &'a Bump) -> Pin<Box<'a, T>>

Constructs a new Pin<Box<T>>. If T does not implement Unpin, then x will be pinned in memory and unable to be moved.

pub fn into_inner(b: Box<'a, T>) -> T

Consumes the Box, returning the wrapped value.

Examples

use bumpalo::{Bump, boxed::Box};

let b = Bump::new();

let hello = Box::new_in("hello".to_owned(), &b);
assert_eq!(Box::into_inner(hello), "hello");

impl<'a, T: ?Sized> Box<'a, T>

pub unsafe fn from_raw(raw: *mut T) -> Self

Constructs a box from a raw pointer.

After calling this function, the raw pointer is owned by the resulting Box. Specifically, the Box destructor will call the destructor of T and free the allocated memory. For this to be safe, the memory must have been allocated in accordance with the memory layout used by Box .

Safety

This function is unsafe because improper use may lead to memory problems. For example, a double-free may occur if the function is called twice on the same raw pointer.

Examples

Recreate a Box which was previously converted to a raw pointer using Box::into_raw:

use bumpalo::{Bump, boxed::Box};

let b = Bump::new();

let x = Box::new_in(5, &b);
let ptr = Box::into_raw(x);
let x = unsafe { Box::from_raw(ptr) }; // Note that new `x`'s lifetime is unbound. It must be bound to the `b` immutable borrow before `b` is reset.

Manually create a Box from scratch by using the bump allocator:

use std::alloc::{alloc, Layout};
use bumpalo::{Bump, boxed::Box};

let b = Bump::new();

unsafe {
    let ptr = b.alloc_layout(Layout::new::<i32>()).as_ptr() as *mut i32;
    *ptr = 5;
    let x = Box::from_raw(ptr); // Note that `x`'s lifetime is unbound. It must be bound to the `b` immutable borrow before `b` is reset.
}

pub fn into_raw(b: Box<'a, T>) -> *mut T

Consumes the Box, returning a wrapped raw pointer.

The pointer will be properly aligned and non-null.

After calling this function, the caller is responsible for the value previously managed by the Box. In particular, the caller should properly destroy T. The easiest way to do this is to convert the raw pointer back into a Box with the Box::from_raw function, allowing the Box destructor to perform the cleanup.

Note: this is an associated function, which means that you have to call it as Box::into_raw(b) instead of b.into_raw(). This is so that there is no conflict with a method on the inner type.

Examples

Converting the raw pointer back into a Box with Box::from_raw for automatic cleanup:

use bumpalo::{Bump, boxed::Box};

let b = Bump::new();

let x = Box::new_in(String::from("Hello"), &b);
let ptr = Box::into_raw(x);
let x = unsafe { Box::from_raw(ptr) }; // Note that new `x`'s lifetime is unbound. It must be bound to the `b` immutable borrow before `b` is reset.

Manual cleanup by explicitly running the destructor:

use std::ptr;
use bumpalo::{Bump, boxed::Box};

let b = Bump::new();

let mut x = Box::new_in(String::from("Hello"), &b);
let p = Box::into_raw(x);
unsafe {
    ptr::drop_in_place(p);
}

pub fn leak(b: Box<'a, T>) -> &'a mut T

Consumes and leaks the Box, returning a mutable reference, &'a mut T. Note that the type T must outlive the chosen lifetime 'a. If the type has only static references, or none at all, then this may be chosen to be 'static.

This function is mainly useful for data that lives for the remainder of the program’s life. Dropping the returned reference will cause a memory leak. If this is not acceptable, the reference should first be wrapped with the Box::from_raw function producing a Box. This Box can then be dropped which will properly destroy T and release the allocated memory.

Note: this is an associated function, which means that you have to call it as Box::leak(b) instead of b.leak(). This is so that there is no conflict with a method on the inner type.

Examples

Simple usage:

use bumpalo::{Bump, boxed::Box};

let b = Bump::new();

let x = Box::new_in(41, &b);
let reference: &mut usize = Box::leak(x);
*reference += 1;
assert_eq!(*reference, 42);

 use bumpalo::{Bump, boxed::Box, vec};

 let b = Bump::new();

 let x = vec![in &b; 1, 2, 3].into_boxed_slice();
 let reference = Box::leak(x);
 reference[0] = 4;
 assert_eq!(*reference, [4, 2, 3]);

impl<'a> Box<'a, dyn Any>

pub fn downcast<T: Any>(self) -> Result<Box<'a, T>, Box<'a, dyn Any>>

Attempt to downcast the box to a concrete type.

Examples

use std::any::Any;

fn print_if_string(value: Box<dyn Any>) {
    if let Ok(string) = value.downcast::<String>() {
        println!("String ({}): {}", string.len(), string);
    }
}

let my_string = "Hello World".to_string();
print_if_string(Box::new(my_string));
print_if_string(Box::new(0i8));

impl<'a> Box<'a, dyn Any + Send>

pub fn downcast<T: Any>(self) -> Result<Box<'a, T>, Box<'a, dyn Any + Send>>

Attempt to downcast the box to a concrete type.

Examples

use std::any::Any;

fn print_if_string(value: Box<dyn Any + Send>) {
    if let Ok(string) = value.downcast::<String>() {
        println!("String ({}): {}", string.len(), string);
    }
}

let my_string = "Hello World".to_string();
print_if_string(Box::new(my_string));
print_if_string(Box::new(0i8));

impl<'a, A> Box<'a, [A]>

pub fn from_iter_in<T: IntoIterator<Item = A>>(iter: T, a: &'a Bump) -> Self

Creates a value from an iterator. This method is an adapted version of FromIterator::from_iter. It cannot be made as that trait implementation given different signature.

Examples

Basic usage:

use bumpalo::{Bump, boxed::Box, vec};

let b = Bump::new();

let five_fives = std::iter::repeat(5).take(5);
let slice = Box::from_iter_in(five_fives, &b);
assert_eq!(vec![in &b; 5, 5, 5, 5, 5], &*slice);