Module bumpalo::collections

Collection types that allocate inside a Bump arena.

Re-exports

pub use self::vec::Vec;

pub use self::string::String;

Modules

string

A UTF-8 encoded, growable string.

vec

A contiguous growable array type with heap-allocated contents, written Vec<'bump, T>.

Enums

CollectionAllocErr

Augments AllocErr with a CapacityOverflow variant.

Traits

CollectIn

Extension trait for iterators, in order to allow allocator-parameterized collections to be constructed more easily.

FromIteratorIn

A trait for types that support being constructed from an iterator, parameterized by an allocator.