Struct bumpalo::collections::vec::Vec

pub struct Vec<'bump, T: 'bump> { /* fields omitted */ }

A contiguous growable array type, written Vec<'bump, T> but pronounced 'vector'.

Examples

use bumpalo::{Bump, collections::Vec};

let b = Bump::new();

let mut vec = Vec::new_in(&b);
vec.push(1);
vec.push(2);

assert_eq!(vec.len(), 2);
assert_eq!(vec[0], 1);

assert_eq!(vec.pop(), Some(2));
assert_eq!(vec.len(), 1);

vec[0] = 7;
assert_eq!(vec[0], 7);

vec.extend([1, 2, 3].iter().cloned());

for x in &vec {
    println!("{}", x);
}
assert_eq!(vec, [7, 1, 2, 3]);

The vec! macro is provided to make initialization more convenient:

use bumpalo::{Bump, collections::Vec};

let b = Bump::new();

let mut vec = bumpalo::vec![in &b; 1, 2, 3];
vec.push(4);
assert_eq!(vec, [1, 2, 3, 4]);

It can also initialize each element of a Vec<'bump, T> with a given value. This may be more efficient than performing allocation and initialization in separate steps, especially when initializing a vector of zeros:

use bumpalo::{Bump, collections::Vec};

let b = Bump::new();

let vec = bumpalo::vec![in &b; 0; 5];
assert_eq!(vec, [0, 0, 0, 0, 0]);

// The following is equivalent, but potentially slower:
let mut vec1 = Vec::with_capacity_in(5, &b);
vec1.resize(5, 0);

Use a Vec<'bump, T> as an efficient stack:

use bumpalo::{Bump, collections::Vec};

let b = Bump::new();

let mut stack = Vec::new_in(&b);

stack.push(1);
stack.push(2);
stack.push(3);

while let Some(top) = stack.pop() {
    // Prints 3, 2, 1
    println!("{}", top);
}

Indexing

The Vec type allows to access values by index, because it implements the Index trait. An example will be more explicit:

use bumpalo::{Bump, collections::Vec};

let b = Bump::new();

let v = bumpalo::vec![in &b; 0, 2, 4, 6];
println!("{}", v[1]); // it will display '2'

However be careful: if you try to access an index which isn't in the Vec, your software will panic! You cannot do this:

use bumpalo::{Bump, collections::Vec};

let b = Bump::new();

let v = bumpalo::vec![in &b; 0, 2, 4, 6];
println!("{}", v[6]); // it will panic!

In conclusion: always check if the index you want to get really exists before doing it.

Slicing

A Vec can be mutable. Slices, on the other hand, are read-only objects. To get a slice, use &. Example:

use bumpalo::{Bump, collections::Vec};

fn read_slice(slice: &[usize]) {
    // ...
}

let b = Bump::new();

let v = bumpalo::vec![in &b; 0, 1];
read_slice(&v);

// ... and that's all!
// you can also do it like this:
let x : &[usize] = &v;

In Rust, it's more common to pass slices as arguments rather than vectors when you just want to provide a read access. The same goes for String and &str.

Capacity and reallocation

The capacity of a vector is the amount of space allocated for any future elements that will be added onto the vector. This is not to be confused with the length of a vector, which specifies the number of actual elements within the vector. If a vector's length exceeds its capacity, its capacity will automatically be increased, but its elements will have to be reallocated.

For example, a vector with capacity 10 and length 0 would be an empty vector with space for 10 more elements. Pushing 10 or fewer elements onto the vector will not change its capacity or cause reallocation to occur. However, if the vector's length is increased to 11, it will have to reallocate, which can be slow. For this reason, it is recommended to use Vec::with_capacity_in whenever possible to specify how big the vector is expected to get.

Guarantees

Due to its incredibly fundamental nature, Vec makes a lot of guarantees about its design. This ensures that it's as low-overhead as possible in the general case, and can be correctly manipulated in primitive ways by unsafe code. Note that these guarantees refer to an unqualified Vec<'bump, T>. If additional type parameters are added (e.g. to support custom allocators), overriding their defaults may change the behavior.

Most fundamentally, Vec is and always will be a (pointer, capacity, length) triplet. No more, no less. The order of these fields is completely unspecified, and you should use the appropriate methods to modify these. The pointer will never be null, so this type is null-pointer-optimized.

However, the pointer may not actually point to allocated memory. In particular, if you construct a Vec with capacity 0 via Vec::new_in, bumpalo::vec![in bump], Vec::with_capacity_in(0), or by calling shrink_to_fit on an empty Vec, it will not allocate memory. Similarly, if you store zero-sized types inside a Vec, it will not allocate space for them. Note that in this case the Vec may not report a capacity of 0. Vec will allocate if and only if mem::size_of::<T>() * capacity() > 0. In general, Vec's allocation details are very subtle — if you intend to allocate memory using a Vec and use it for something else (either to pass to unsafe code, or to build your own memory-backed collection), be sure to deallocate this memory by using from_raw_parts to recover the Vec and then dropping it.

If a Vec has allocated memory, then the memory it points to is on the heap (as defined by the allocator Rust is configured to use by default), and its pointer points to len initialized, contiguous elements in order (what you would see if you coerced it to a slice), followed by capacity-len logically uninitialized, contiguous elements.

Vec will never perform a "small optimization" where elements are actually stored on the stack for two reasons:

It would make it more difficult for unsafe code to correctly manipulate a Vec. The contents of a Vec wouldn't have a stable address if it were only moved, and it would be more difficult to determine if a Vec had actually allocated memory.

It would penalize the general case, incurring an additional branch on every access.

Vec will never automatically shrink itself, even if completely empty. This ensures no unnecessary allocations or deallocations occur. Emptying a Vec and then filling it back up to the same len should incur no calls to the allocator. If you wish to free up unused memory, use shrink_to_fit.

push and insert will never (re)allocate if the reported capacity is sufficient. push and insert will (re)allocate if len==capacity. That is, the reported capacity is completely accurate, and can be relied on. It can even be used to manually free the memory allocated by a Vec if desired. Bulk insertion methods may reallocate, even when not necessary.

Vec does not guarantee any particular growth strategy when reallocating when full, nor when reserve is called. The current strategy is basic and it may prove desirable to use a non-constant growth factor. Whatever strategy is used will of course guarantee O(1) amortized push.

bumpalo::vec![in bump; x; n], bumpalo::vec![in bump; a, b, c, d], and Vec::with_capacity_in(n), will all produce a Vec with exactly the requested capacity. If len==capacity, (as is the case for the vec! macro), then a Vec<'bump, T> can be converted to and from a Box<[T]> without reallocating or moving the elements.

Vec will not specifically overwrite any data that is removed from it, but also won't specifically preserve it. Its uninitialized memory is scratch space that it may use however it wants. It will generally just do whatever is most efficient or otherwise easy to implement. Do not rely on removed data to be erased for security purposes. Even if you drop a Vec, its buffer may simply be reused by another Vec. Even if you zero a Vec's memory first, that may not actually happen because the optimizer does not consider this a side-effect that must be preserved. There is one case which we will not break, however: using unsafe code to write to the excess capacity, and then increasing the length to match, is always valid.

Vec does not currently guarantee the order in which elements are dropped. The order has changed in the past and may change again.

Implementations

impl<'bump, T: 'bump> Vec<'bump, T>

pub fn new_in(bump: &'bump Bump) -> Vec<'bump, T>

Constructs a new, empty Vec<'bump, T>.

The vector will not allocate until elements are pushed onto it.

Examples

use bumpalo::{Bump, collections::Vec};

let b = Bump::new();
let mut vec: Vec<i32> = Vec::new_in(&b);

pub fn with_capacity_in(capacity: usize, bump: &'bump Bump) -> Vec<'bump, T>

Constructs a new, empty Vec<'bump, T> with the specified capacity.

The vector will be able to hold exactly capacity elements without reallocating. If capacity is 0, the vector will not allocate.

It is important to note that although the returned vector has the capacity specified, the vector will have a zero length. For an explanation of the difference between length and capacity, see Capacity and reallocation.

Examples

use bumpalo::{Bump, collections::Vec};

let b = Bump::new();

let mut vec = Vec::with_capacity_in(10, &b);

// The vector contains no items, even though it has capacity for more
assert_eq!(vec.len(), 0);

// These are all done without reallocating...
for i in 0..10 {
    vec.push(i);
}

// ...but this may make the vector reallocate
vec.push(11);

pub fn from_iter_in<I: IntoIterator<Item = T>>(
iter: I,
bump: &'bump Bump
) -> Vec<'bump, T>

Construct a new Vec from the given iterator's items.

Examples

use bumpalo::{Bump, collections::Vec};
use std::iter;

let b = Bump::new();
let v = Vec::from_iter_in(iter::repeat(7).take(3), &b);
assert_eq!(v, [7, 7, 7]);

pub unsafe fn from_raw_parts_in(
ptr: *mut T,
length: usize,
capacity: usize,
bump: &'bump Bump
) -> Vec<'bump, T>

Creates a Vec<'bump, T> directly from the raw components of another vector.

Safety

This is highly unsafe, due to the number of invariants that aren't checked:

ptr needs to have been previously allocated via String/Vec<'bump, T> (at least, it's highly likely to be incorrect if it wasn't).

ptr's T needs to have the same size and alignment as it was allocated with.

length needs to be less than or equal to capacity.

capacity needs to be the capacity that the pointer was allocated with.

Violating these may cause problems like corrupting the allocator's internal data structures. For example it is not safe to build a Vec<u8> from a pointer to a C char array and a size_t.

The ownership of ptr is effectively transferred to the Vec<'bump, T> which may then deallocate, reallocate or change the contents of memory pointed to by the pointer at will. Ensure that nothing else uses the pointer after calling this function.

Examples

use bumpalo::{Bump, collections::Vec};

use std::ptr;
use std::mem;

fn main() {
    let b = Bump::new();

    let mut v = bumpalo::vec![in &b; 1, 2, 3];

    // Pull out the various important pieces of information about `v`
    let p = v.as_mut_ptr();
    let len = v.len();
    let cap = v.capacity();

    unsafe {
        // Cast `v` into the void: no destructor run, so we are in
        // complete control of the allocation to which `p` points.
        mem::forget(v);

        // Overwrite memory with 4, 5, 6
        for i in 0..len as isize {
            ptr::write(p.offset(i), 4 + i);
        }

        // Put everything back together into a Vec
        let rebuilt = Vec::from_raw_parts_in(p, len, cap, &b);
        assert_eq!(rebuilt, [4, 5, 6]);
    }
}

pub fn capacity(&self) -> usize

Returns the number of elements the vector can hold without reallocating.

Examples

use bumpalo::{Bump, collections::Vec};

let b = Bump::new();
let vec: Vec<i32> = Vec::with_capacity_in(10, &b);
assert_eq!(vec.capacity(), 10);

pub fn reserve(&mut self, additional: usize)

Reserves capacity for at least additional more elements to be inserted in the given Vec<'bump, T>. The collection may reserve more space to avoid frequent reallocations. After calling reserve, capacity will be greater than or equal to self.len() + additional. Does nothing if capacity is already sufficient.

Panics

Panics if the new capacity overflows usize.

Examples

use bumpalo::{Bump, collections::Vec};

let b = Bump::new();
let mut vec = bumpalo::vec![in &b; 1];
vec.reserve(10);
assert!(vec.capacity() >= 11);

pub fn reserve_exact(&mut self, additional: usize)

Reserves the minimum capacity for exactly additional more elements to be inserted in the given Vec<'bump, T>. After calling reserve_exact, capacity will be greater than or equal to self.len() + additional. Does nothing if the capacity is already sufficient.

Note that the allocator may give the collection more space than it requests. Therefore capacity can not be relied upon to be precisely minimal. Prefer reserve if future insertions are expected.

Panics

Panics if the new capacity overflows usize.

Examples

use bumpalo::{Bump, collections::Vec};

let b = Bump::new();
let mut vec = bumpalo::vec![in &b; 1];
vec.reserve_exact(10);
assert!(vec.capacity() >= 11);

pub fn shrink_to_fit(&mut self)

Shrinks the capacity of the vector as much as possible.

It will drop down as close as possible to the length but the allocator may still inform the vector that there is space for a few more elements.

Examples

use bumpalo::{Bump, collections::Vec};

let b = Bump::new();

let mut vec = Vec::with_capacity_in(10, &b);
vec.extend([1, 2, 3].iter().cloned());
assert_eq!(vec.capacity(), 10);
vec.shrink_to_fit();
assert!(vec.capacity() >= 3);

pub fn into_bump_slice(self) -> &'bump [T]

Converts the vector into &'bump [T].

Examples

use bumpalo::{Bump, collections::Vec};

let b = Bump::new();
let v = bumpalo::vec![in &b; 1, 2, 3];

let slice = v.into_bump_slice();
assert_eq!(slice, [1, 2, 3]);

pub fn truncate(&mut self, len: usize)

Shortens the vector, keeping the first len elements and dropping the rest.

If len is greater than the vector's current length, this has no effect.

The drain method can emulate truncate, but causes the excess elements to be returned instead of dropped.

Note that this method has no effect on the allocated capacity of the vector.

Examples

Truncating a five element vector to two elements:

use bumpalo::{Bump, collections::Vec};

let b = Bump::new();

let mut vec = bumpalo::vec![in &b; 1, 2, 3, 4, 5];
vec.truncate(2);
assert_eq!(vec, [1, 2]);

No truncation occurs when len is greater than the vector's current length:

use bumpalo::{Bump, collections::Vec};

let b = Bump::new();

let mut vec = bumpalo::vec![in &b; 1, 2, 3];
vec.truncate(8);
assert_eq!(vec, [1, 2, 3]);

Truncating when len == 0 is equivalent to calling the clear method.

use bumpalo::{Bump, collections::Vec};

let b = Bump::new();

let mut vec = bumpalo::vec![in &b; 1, 2, 3];
vec.truncate(0);
assert_eq!(vec, []);

pub fn as_slice(&self) -> &[T]

Extracts a slice containing the entire vector.

Equivalent to &s[..].

Examples

use bumpalo::{Bump, collections::Vec};
use std::io::{self, Write};

let b = Bump::new();

let buffer = bumpalo::vec![in &b; 1, 2, 3, 5, 8];
io::sink().write(buffer.as_slice()).unwrap();

pub fn as_mut_slice(&mut self) -> &mut [T]

Extracts a mutable slice of the entire vector.

Equivalent to &mut s[..].

Examples

use bumpalo::{Bump, collections::Vec};
use std::io::{self, Read};

let b = Bump::new();
let mut buffer = bumpalo::vec![in &b; 0; 3];
io::repeat(0b101).read_exact(buffer.as_mut_slice()).unwrap();

pub unsafe fn set_len(&mut self, len: usize)

Sets the length of a vector.

This will explicitly set the size of the vector, without actually modifying its buffers, so it is up to the caller to ensure that the vector is actually the specified size.

Examples

use bumpalo::{Bump, collections::Vec};

use std::ptr;

let b = Bump::new();

let mut vec = bumpalo::vec![in &b; 'r', 'u', 's', 't'];

unsafe {
    ptr::drop_in_place(&mut vec[3]);
    vec.set_len(3);
}
assert_eq!(vec, ['r', 'u', 's']);

In this example, there is a memory leak since the memory locations owned by the inner vectors were not freed prior to the set_len call:

use bumpalo::{Bump, collections::Vec};

let b = Bump::new();

let mut vec = bumpalo::vec![in &b;
                            bumpalo::vec![in &b; 1, 0, 0],
                            bumpalo::vec![in &b; 0, 1, 0],
                            bumpalo::vec![in &b; 0, 0, 1]];
unsafe {
    vec.set_len(0);
}

In this example, the vector gets expanded from zero to four items without any memory allocations occurring, resulting in vector values of unallocated memory:

use bumpalo::{Bump, collections::Vec};

let b = Bump::new();

let mut vec: Vec<char> = Vec::new_in(&b);

unsafe {
    vec.set_len(4);
}

pub fn swap_remove(&mut self, index: usize) -> T

Removes an element from the vector and returns it.

The removed element is replaced by the last element of the vector.

This does not preserve ordering, but is O(1).

Panics

Panics if index is out of bounds.

Examples

use bumpalo::{Bump, collections::Vec};

let b = Bump::new();

let mut v = bumpalo::vec![in &b; "foo", "bar", "baz", "qux"];

assert_eq!(v.swap_remove(1), "bar");
assert_eq!(v, ["foo", "qux", "baz"]);

assert_eq!(v.swap_remove(0), "foo");
assert_eq!(v, ["baz", "qux"]);

pub fn insert(&mut self, index: usize, element: T)

Inserts an element at position index within the vector, shifting all elements after it to the right.

Panics

Panics if index > len.

Examples

use bumpalo::{Bump, collections::Vec};

let b = Bump::new();

let mut vec = bumpalo::vec![in &b; 1, 2, 3];
vec.insert(1, 4);
assert_eq!(vec, [1, 4, 2, 3]);
vec.insert(4, 5);
assert_eq!(vec, [1, 4, 2, 3, 5]);

pub fn remove(&mut self, index: usize) -> T

Removes and returns the element at position index within the vector, shifting all elements after it to the left.

Panics

Panics if index is out of bounds.

Examples

use bumpalo::{Bump, collections::Vec};

let b = Bump::new();

let mut v = bumpalo::vec![in &b; 1, 2, 3];
assert_eq!(v.remove(1), 2);
assert_eq!(v, [1, 3]);

pub fn retain<F>(&mut self, f: F) where
F: FnMut(&T) -> bool,

Retains only the elements specified by the predicate.

In other words, remove all elements e such that f(&e) returns false. This method operates in place and preserves the order of the retained elements.

Examples

use bumpalo::{Bump, collections::Vec};

let b = Bump::new();

let mut vec = bumpalo::vec![in &b; 1, 2, 3, 4];
vec.retain(|&x| x%2 == 0);
assert_eq!(vec, [2, 4]);

pub fn dedup_by_key<F, K>(&mut self, key: F) where
F: FnMut(&mut T) -> K,
K: PartialEq,

Removes all but the first of consecutive elements in the vector that resolve to the same key.

If the vector is sorted, this removes all duplicates.

Examples

use bumpalo::{Bump, collections::Vec};

let b = Bump::new();

let mut vec = bumpalo::vec![in &b; 10, 20, 21, 30, 20];

vec.dedup_by_key(|i| *i / 10);

assert_eq!(vec, [10, 20, 30, 20]);

pub fn dedup_by<F>(&mut self, same_bucket: F) where
F: FnMut(&mut T, &mut T) -> bool,

Removes all but the first of consecutive elements in the vector satisfying a given equality relation.

The same_bucket function is passed references to two elements from the vector and must determine if the elements compare equal. The elements are passed in opposite order from their order in the slice, so if same_bucket(a, b) returns true, a is removed.

If the vector is sorted, this removes all duplicates.

Examples

use bumpalo::{Bump, collections::Vec};

let b = Bump::new();

let mut vec = bumpalo::vec![in &b; "foo", "bar", "Bar", "baz", "bar"];

vec.dedup_by(|a, b| a.eq_ignore_ascii_case(b));

assert_eq!(vec, ["foo", "bar", "baz", "bar"]);

pub fn push(&mut self, value: T)

Appends an element to the back of a collection.

Panics

Panics if the number of elements in the vector overflows a usize.

Examples

use bumpalo::{Bump, collections::Vec};

let b = Bump::new();

let mut vec = bumpalo::vec![in &b; 1, 2];
vec.push(3);
assert_eq!(vec, [1, 2, 3]);

pub fn pop(&mut self) -> Option<T>

Removes the last element from a vector and returns it, or None if it is empty.

Examples

use bumpalo::{Bump, collections::Vec};

let b = Bump::new();

let mut vec = bumpalo::vec![in &b; 1, 2, 3];
assert_eq!(vec.pop(), Some(3));
assert_eq!(vec, [1, 2]);

pub fn append(&mut self, other: &mut Self)

Moves all the elements of other into Self, leaving other empty.

Panics

Panics if the number of elements in the vector overflows a usize.

Examples

use bumpalo::{Bump, collections::Vec};

let b = Bump::new();

let mut vec = bumpalo::vec![in &b; 1, 2, 3];
let mut vec2 = bumpalo::vec![in &b; 4, 5, 6];
vec.append(&mut vec2);
assert_eq!(vec, [1, 2, 3, 4, 5, 6]);
assert_eq!(vec2, []);

pub fn drain<R>(&mut self, range: R) -> Drain<T> where
R: RangeBounds<usize>,

Creates a draining iterator that removes the specified range in the vector and yields the removed items.

Note 1: The element range is removed even if the iterator is only partially consumed or not consumed at all.

Note 2: It is unspecified how many elements are removed from the vector if the Drain value is leaked.

Panics

Panics if the starting point is greater than the end point or if the end point is greater than the length of the vector.

Examples

use bumpalo::{Bump, collections::Vec};

let b = Bump::new();

let mut v = bumpalo::vec![in &b; 1, 2, 3];

let mut u: Vec<_> = Vec::new_in(&b);
u.extend(v.drain(1..));

assert_eq!(v, &[1]);
assert_eq!(u, &[2, 3]);

// A full range clears the vector
v.drain(..);
assert_eq!(v, &[]);

pub fn clear(&mut self)

Clears the vector, removing all values.

Note that this method has no effect on the allocated capacity of the vector.

Examples

use bumpalo::{Bump, collections::Vec};

let b = Bump::new();

let mut v = bumpalo::vec![in &b; 1, 2, 3];

v.clear();

assert!(v.is_empty());

pub fn len(&self) -> usize

Returns the number of elements in the vector, also referred to as its 'length'.

Examples

use bumpalo::{Bump, collections::Vec};

let b = Bump::new();

let a = bumpalo::vec![in &b; 1, 2, 3];
assert_eq!(a.len(), 3);

pub fn is_empty(&self) -> bool

Returns true if the vector contains no elements.

Examples

use bumpalo::{Bump, collections::Vec};

let b = Bump::new();

let mut v = Vec::new_in(&b);
assert!(v.is_empty());

v.push(1);
assert!(!v.is_empty());

pub fn split_off(&mut self, at: usize) -> Self

Splits the collection into two at the given index.

Returns a newly allocated Self. self contains elements [0, at), and the returned Self contains elements [at, len).

Note that the capacity of self does not change.

Panics

Panics if at > len.

Examples

use bumpalo::{Bump, collections::Vec};

let b = Bump::new();

let mut vec = bumpalo::vec![in &b; 1,2,3];
let vec2 = vec.split_off(1);
assert_eq!(vec, [1]);
assert_eq!(vec2, [2, 3]);

impl<'bump, T: 'bump + Clone> Vec<'bump, T>

pub fn resize(&mut self, new_len: usize, value: T)

Resizes the Vec in-place so that len is equal to new_len.

If new_len is greater than len, the Vec is extended by the difference, with each additional slot filled with value. If new_len is less than len, the Vec is simply truncated.

This method requires Clone to be able clone the passed value. If you need more flexibility (or want to rely on Default instead of Clone), use resize_with.

Examples

use bumpalo::{Bump, collections::Vec};

let b = Bump::new();

let mut vec = bumpalo::vec![in &b; "hello"];
vec.resize(3, "world");
assert_eq!(vec, ["hello", "world", "world"]);

let mut vec = bumpalo::vec![in &b; 1, 2, 3, 4];
vec.resize(2, 0);
assert_eq!(vec, [1, 2]);

pub fn extend_from_slice(&mut self, other: &[T])

Clones and appends all elements in a slice to the Vec.

Iterates over the slice other, clones each element, and then appends it to this Vec. The other vector is traversed in-order.

Note that this function is same as extend except that it is specialized to work with slices instead. If and when Rust gets specialization this function will likely be deprecated (but still available).

Examples

use bumpalo::{Bump, collections::Vec};

let b = Bump::new();

let mut vec = bumpalo::vec![in &b; 1];
vec.extend_from_slice(&[2, 3, 4]);
assert_eq!(vec, [1, 2, 3, 4]);

impl<'bump, T: 'bump + PartialEq> Vec<'bump, T>

pub fn dedup(&mut self)

Removes consecutive repeated elements in the vector according to the PartialEq trait implementation.

If the vector is sorted, this removes all duplicates.

Examples

use bumpalo::{Bump, collections::Vec};

let b = Bump::new();

let mut vec = bumpalo::vec![in &b; 1, 2, 2, 3, 2];

vec.dedup();

assert_eq!(vec, [1, 2, 3, 2]);

impl<'bump, T: 'bump> Vec<'bump, T>

pub fn splice<R, I>(&mut self, range: R, replace_with: I) -> Splice<I::IntoIter> where
R: RangeBounds<usize>,
I: IntoIterator<Item = T>,

Creates a splicing iterator that replaces the specified range in the vector with the given replace_with iterator and yields the removed items. replace_with does not need to be the same length as range.

Note 1: The element range is removed even if the iterator is not consumed until the end.

Note 2: It is unspecified how many elements are removed from the vector, if the Splice value is leaked.

Note 3: The input iterator replace_with is only consumed when the Splice value is dropped.

Note 4: This is optimal if:

The tail (elements in the vector after range) is empty,

or replace_with yields fewer elements than range’s length

or the lower bound of its size_hint() is exact.

Otherwise, a temporary vector is allocated and the tail is moved twice.

Panics

Panics if the starting point is greater than the end point or if the end point is greater than the length of the vector.

Examples

use bumpalo::{Bump, collections::Vec};

let b = Bump::new();

let mut v = bumpalo::vec![in &b; 1, 2, 3];
let new = [7, 8];
let u: Vec<_> = Vec::from_iter_in(v.splice(..2, new.iter().cloned()), &b);
assert_eq!(v, &[7, 8, 3]);
assert_eq!(u, &[1, 2]);